TCPCI: Update RP value selection has to set the register
The tcpci_tcpm_select_rp_value was only caching the
value to be used in case we set other fields and had
to recreate.  This function did not actually set the
Rp value in the Role Control register. This made
PD3 collision detection work incorrectly.

This was seen on ThinkPad Gen2 Dock connecting to
trembyle.  On first attach the DUT was not correctly
reflecting SinkTxNG on Rp and the dock detected
this after the PR_Swap was sent and performed
Soft and Hard resets

BUG=b:157755939
BRANCH=none
TEST=Verify Rp tracks SinkTxNG

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I5e2cd1bd2c7fffa06f46e1d7a77868f49454c591