Update patch set 3
Patch Set 3: Verified+1

I don't want to add logic into the Charge Manager to determine the correct pull to set for a set_cc call so added a new function that will just set the rp portion of the Role Control and leave the rest of the register as it was. This will be similar to calling set_cc but in a safer manner for the Charge Manager. The other calls to select_rp_value will remain exactly as they were.

Just making this one set_rp actually change the value in the Role Control got rid of my Dock Reset issue.

Patch-set: 3
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1