Trogdor: Skip resetting the power supply, if VBUS not enabled
The TCPMv2 abuses calling the pd_power_supply_reset() call, even
VBUS is not enabled.

This happens on the CHIPSET beginining the transition from S0 to S3.
A hook is triggered to set the dual role flag from toggle-on to
toggle-off. If the port has nothing attached, its Type-C state is
changed from DRPAutoToggle to Unattached.SNK. The Unattached.SNK is
a substate of CC_RD. The entry of CC_RD calls pd_power_supply_reset()
to reset power.

The pd_power_supply_reset() will then wake AP up and notify AP about
the power info change, even there is no power change. This prevents
AP going to S3.

Before the TCPMv2 has a proper fix for this abuse call, let's skip
the function if VBUS is not enbled.

BRANCH=None
BUG=b:157685747
TEST=Modified AP_SUSPEND to high, checked the following event not
triggered to wake AP up.
[event set 0x0000000000200000]

Change-Id: I6b544cc4a57f23869d9cfb0af822b1281bebba13
Signed-off-by: Wai-Hong Tam <waihong@google.com>