Dummy commit to bump version
Working around b/149333727 and making sure that we have an MP-signed
version of the firmware that does not have the same version as the
Pre-MP signed version to avoid confusion (especially in the factory).

The build does not let us use an empty commit, so adding a dummy commit
and reverting it: https://crbug.com/1090423.

BRANCH=none
BUG=b:154365840, b:149333727, chromium:1090423
TEST=make buildall -j

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I3413fbffe6cfb21713df5847cca6bdf8cf187f84