Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Tom, do you see any obvious problems in not resetting backup domain when configuring RTC clock? I'll build bloonchipper now and verify that fingerprint functionality is not broken by this.
> 
> I'm reviewing the datasheet and looking at other chips to double-check.

This looks like it's ok. According to "6.1.3 Backup domain reset" in RM0402, the backup domain is reset when "VDD or VBAT power on, if both supplies have previously been powered off.", so it will be reset at the initial power on.

I think the only question is the difference in behavior for the remaining non-scratchpad registers. With this change, these will no longer be cleared, but it looks like in the case of both the reset flags and the panic flags, those will be reset during init after they are read:

https://chromium.googlesource.com/chromiumos/platform/ec/+/master/chip/stm32/system.c#96
https://chromium.googlesource.com/chromiumos/platform/ec/+/master/chip/stm32/system.c#308

The main difference I see with this implementation versus the reset flags and panic flag implementation is that the reset and panic flags will not persist across multiple resets since we explicitly read them out and clear them, but with this change the scratchpad register would. I think that's ok since we're only using the scratchpad in the tests.

Patch-set: 1
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2