Update patch set 1
Patch Set 1:

Yicheng, I went ahead and cherry-picked this one since it will take a while for the other change to go through the CQ for master.

We can do a merge to the release branch tomorrow (like we've been doing), but this will make sure that this change lands for the next build.

Patch-set: 1
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>