hooks: Introduce HOOK_CHIPSET_POST_SHUTDOWN
A new hook HOOK_CHIPSET_POST_SHUTDOWN is introduced, which are called
from the chipset task, while the system has already shut down and all
the suspend rails are already off.

It will be used for executing pending EC reboot at the chipset shutdown.
The EC reboot should be executed when the chipset is completely off.

BRANCH=None
BUG=b:156981868
TEST=Built all boards.

Change-Id: I12f26957e46a1bb34ef079f127b0bddd133cd4e7
Signed-off-by: Wai-Hong Tam <waihong@google.com>