system: Do pending reboot when the chipset is completely off
The pending reboot command is currently triggered in the
CHIPSET_SHUTDOWN hook.

The CHIPSET_SHUTDOWN hook is triggered before the power rails
are removed. The hook is executed in the CHIPSET task context,
meaning the chipset is still ON.

Should wait for the chipset is completely off. All the power rails
are turned off. So change the hook to the CHIPSET_SHUTDOWN_COMPLETE.

Also make it the lowest priority. So other hooks have chances to
be executed before EC reboot.

BRANCH=None
BUG=b:156981868
TEST=Ran "ectool reboot_ec cold at-shutdown" and "shutdown -H now".
[45.061296 power state 8 = S3->S5, in 0x0005]
[45.068785 Base Accel ODR: 0 - roundup 0 from config0 [AP 0]]
[45.070271 Base Accel ODR: 0 - roundup 0 from config0 [AP 0]]
[45.076021 Gyro ODR: 0 - roundup 0 from config 0 [AP0]]
[45.078307 PD:S3->S5]
[45.078877 Gyro ODR: 0 - roundup 0 from config 0 [AP0]]
[45.080055 set_pmic_pwron(0)]
[45.150585 set_system_power(0)]
[45.161457 power shutdown complete]
Reboot at shutdown: 4

Change-Id: Ic8258cbf35fccff633d0c9ccd5d09faefc594855
Signed-off-by: Wai-Hong Tam <waihong@google.com>