Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is it really necessary to rename this to PRE_SHUTDOWN? I think POST_SHUTDOWN can stand on its own. CHIPSET_SHUTDOWN means the chipset *is* shutting down. The symmetry of pre/post shutdown doesn't really fit in my opinion; to me, PRE_SHUTDOWN means the chipset is *about to* shutdown.

This is the definition in hooks.h:
```
System is shutting down.  All suspend rails are still on.
```

The wording of "is shutting" may be confusing. The next sentence "All suspend rails are still on" explains it is called before the suspend rails are shut down. The 'pre' makes sense, at least in this definition.

Patch-set: 2
Reviewer: Gerrit User 1000407 <1000407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>