NPCX7M7WC: correct image size to avoid overflowing WP_RO
The -7M7WC chip has half as much flash as the -7M7WB but the same amount
of RAM, which causes the EC_RO section in the output FMAP to be larger
than the actual read-only section of flash because code RAM is not more
than twice the size of the on-chip flash, which is required because RO
also contains the bootloader header.

Update the definitions for this chip to generate a smaller image that
still fits in Flash after the RO header is added. Because this breaks
assumptions about used memory being equal to available memory, remove
the RAM size checks and reorganize the chip blocks to define code and
data memory in the same block to obviate any need for RAM size checks.

BUG=b:158052612
TEST=Verify flash map is not broken: dump_fmap -hh build/dalboz/ec.bin,
     where prior to this change it complains that sections overlap:
         0x00000000:0x0003f03f EC_RO
         0x00000040:0x0003f03f FR_MAIN
         0x00000184:0x000001a3 RO_FRID
         0x0002bf40:0x0002c09d FMAP
         0x00000000:0x0003ffff WP_RO
         0x00040000:0x0007efff EC_RW
         0x00040144:0x00040163 RW_FWID

     Flashing the new image to a Dalboz boots and jumps to RW correctly.
BRANCH=None

Change-Id: I2ac28a7f973c4ae715828687edaad5f0110a2950
Signed-off-by: Peter Marheine <pmarheine@chromium.org>