Update patch set 1
Patch Set 1:

> Patch Set 1: Verified+1
> 
> I think for now this charger is only used on Volteer and Zork. It would be good if you could try this out Denis, I don't know how this would create new issues though.

I do not have a current probe, so I can't really say I was able to test the change.  I did, however, cherry-pick in the change and ran through a few attach/unattach sequences with a charger and a powered hub. Both appeared to be functioning correctly.  If you ask klug@ to verify this, he would have the ability to do so on a trembyle board.

Patch-set: 1