Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Verified+1
> > 
> > I think for now this charger is only used on Volteer and Zork. It would be good if you could try this out Denis, I don't know how this would create new issues though.
> 
> I do not have a current probe, so I can't really say I was able to test the change.  I did, however, cherry-pick in the change and ran through a few attach/unattach sequences with a charger and a powered hub. Both appeared to be functioning correctly.  If you ask klug@ to verify this, he would have the ability to do so on a trembyle board.

I don't think its a big deal that you don't have a current probe. This can only improve the current spiking behavior, I would expect that it was there before and is now gone. If Kevin has a chance to check on a Zork that would be great, if not I'm not worried about it as long as you did a few adapter insertions.

Patch-set: 1
CC: Gerrit User 1364414 <1364414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1347775 <1347775@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>