Update patch set 1
Patch Set 1:

The actual part number is NPCX797WC, and it may be better to reference this code in the CL description 
The naming of the variant is, err, interesting...
It is described in config_chip-npcx7.h but doesn't seem to be followed closely i.e
most of them have M instead of 9 (maybe that's a wildcard that includes both 8 & 9).
Not sure why, but it's incredibly confusing. To my mind, the variant should be the part number i.e
NPCX797WC not NPCX7M7WC.
It would be nice to correct all the others :-) but I would settle just for this variant.

Patch-set: 1
Reviewer: Gerrit User 1281040 <1281040@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>