usb_mux: ps8822: Initial version of driver
This CL adds the driver for the PS8822 usb demux. Currently, the
driver only supports the expected usb mux driver methods. It is
expected that any EQ setting functions would be board specific, though
more register definitions would need to be added to the .h file.

BUG=b:175660576
BRANCH=None
TEST=Connect on host port and then read back mode register to confirm
that it's been set as expected.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ic425ffe41f4653cee7f5500b081d7b5f6ad1a3df