util/flash_ec: Remove ec name check
This name check renders this script useless if the ec directory is not
specifically name "ec". This is not totally a valid assumption, plus
I can't quite determine what this check helps.

If $0 provides a bogus value and this check is triggered, flash_ec
simply provides the following error:

ERROR: no EC image found : build one or specify one.

This error would also be emitted if $0 was bogus without this check.

BRANCH=none
BUG=none
TEST=cd ec
     ./util/flash_ec --board=bloonchipper
TEST=cd ec2
     ./util/flash_ec --board=bloonchipper

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: Ic8a4c6f517d1945397d43a1201001179d2829bcb