Update patch set 3
Patch Set 3:

When attaching the dock with a dongle on the other port, I was able to get past the reset hit that I was hitting before. So this is getting closer. I am not able to complete the attach.

Patch-set: 3
CC: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>