Update patch set 2
Patch Set 2:

No negative feedback, I just wanted to know the reasoning and intention. 400 bytes is somewhat significant and may not fit on every board (e.g. eve). However, as stated in the bug, if the intention is that every baseboard will have this, then it makes sense to enable it by default. I'm therefore assuming that if a board needs some more space, they should be ok to disable this. Hopefully one of their peer boards would still have it enabled.

Patch-set: 2