Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I didn't work on Octopus, so I'll defer to Jett.

My mistake. Sorry to you and Jett.

Hi Jett,
'chargen' command shall be used to test cr50 UART bridging function. Octopus is an ideal baseboard for this purpose since it has npcx and ite ec types. This command occupies 420 bytes or so, and I've checked that still there are lots of free space in flash memory. wdyt?

Patch-set: 5