tcpmv2: process TCH before and after prl_tx
We need to process the TX chunking state machine both before and after
the prl_tx state machine.

Processing before allows us to send outgoing chunked messages on this
iteration.

Processing after allows us to send the TX_COMPLETE status to the PE
layer on the same iteration the PRL gets its notification.

Without this change, if the TCPC fires its alert saying we finished the
TX and we also receive a message (in response to that TX), then we would
get a RX message/event before the TX_COMPLETE notification was sent in
the PE layer.

BRANCH=none
BUG=b:158248741
TEST=Verify the RX and TX messages are happening in the same
statemachine cycle instead of getting the RX event first.

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: I6fd2ece2326f8dc3cee85c94baae09466d673e87