waddledoo: Use TCPCI cached VBUS presence
This commit changes the mechanism in which waddledoo checks VBUS
presence.  Instead of going out to talk to each TCPC, now waddledoo
will simply use the cached VBUS presence within the TCPCI driver.
This was causing a deadlock leading to watchdog resets when TCPC low
power mode was enabled.

BUG=b:158218613
BRANCH=None
TEST=Enable TCPC low power mode, build and flash waddledoo, verify
that the DUT does not watchdog reset.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ifee42979246bb378e9af226123164b10b2df146e