raa489000: Don't write to reserved registers
In the early revision of silicon for the RAA489000, the EC had to make
a number of writes to certain registers.  However, in the latest
datasheets, a lot of the registers are now reserved.  Therefore, we
shouldn't be writing to them anymore.

This commit simply restricts those register writes to the older
version of silicon where the registers were still valid.

BUG=b:158611239
BRANCH=None
TEST=Build and flash on waddledoo with newer silicon, verify that TCPC
is still working and were are able to DRP toggle, respond to PD
messages, negotiate higher contracts, enter DP alt mode.  Verify that
both orientations still work.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I33e3fafa50f37bb1de5bc31fb0618867e5e47a53