TCPMv2: Handle message discard in the PE
This commit creates a reporting function for the PRL to let the PE know
when messages have been discarded.  The PE can return to source or sink
ready to handle the discard if the message was the first of an AMS.  To
manage this, a new flag was added and a shared parent state which will
set this flag and also run the SenderResponseTimer in a unified way.

BRANCH=None
BUG=b:158248741,b:157228506
TEST=on waddledoo, StarTeck dock correctly moves back to ready on
discard of discovery VDM

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I61b3e96e1d14634dcd5edda4e133968ace769b21