TCPMv2: Handle message discard in the PE
This commit creates a reporting function for the PRL to let the PE know
when messages have been discarded.

States will have to handle message discard internally since the response
depends on the portion of the AMS sequence they're in.  For some states
which begin AMSes, a new shared parent has been added which runs the
sender response timer and checks discard for the child.

The new parent state saves an average of 124 bytes of flash space.

BRANCH=None
BUG=b:158248741,b:157228506
TEST=on waddledoo, Dell dock no longer soft resets during power role
swap

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I61b3e96e1d14634dcd5edda4e133968ace769b21