Update patch set 1
Patch Set 1:

Just to confirm after checking with Sam, for the version 2 boards, some had IDs allocated but are not built or used, and go/puff-skus should have the complete list for board version 2.
I do want to keep the list to just the boards that require it.

Patch-set: 1
Label: Commit-Queue=0