Update patch set 1
Patch Set 1: Verified+1

This works for a non-pd charger, but on my waddledoo, I'm not seeing the Type-C supplier getting set when using a Suzyqable. It'd be great if others could let me know if they're seeing that or not and if the behaviour differs with or without this patch.

Patch-set: 1
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1