PB: Clear AP_IDLE for recovery boot
Currently, if CONFIG_POWER_BUTTON_INIT_IDLE is defined,
EC_RESET_FLAG_AP_IDLE is set when the system enters S5. This is for
restoring the power state after power loss or blackout.

This can cause a recovery boot to be blocked depending on the timing
that the hardware resets the EC.

This patch makes EC clear AP_IDLE flag when a manual recovery request
is detected so that AP can be unconditionally booted in the recovery
mode.

BUG=b:158802935, b:37536389
BRANCH=none
TEST=Shutdown DUT. Press power + recovery. DUT enters recovery mode.
TEST=Shutdown DUT. Replug AC. DUT stays off.
TEST=Boot DUT to OS. Replug AC. DUT boots back to OS.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Iec1fd51cb0fa49e088f3b2489c5e94c8a4777a9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2242664
Reviewed-by: Andrew McRae <amcrae@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
(cherry picked from commit e6a74a364207fa0203294f6da3570baf067ff9f2)