PB: Clear AP_IDLE for recovery boot
Currently, if CONFIG_POWER_BUTTON_INIT_IDLE is defined,
EC_RESET_FLAG_AP_IDLE is set when the system enters S5. This is for
restoring the power state after power loss or blackout.

This can cause a recovery boot to be blocked depending on the timing
that the hardware resets the EC.

BUG=b:158802935
BRANCH=none
TEST=Verified recovery button can boot Puff immediately from G3.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Iec1fd51cb0fa49e088f3b2489c5e94c8a4777a9b