Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is it ready to be reviewed? And looks like create_initial_ec_image.sh is up to 1.0.3 already?

In this case, it is not necessary to re-create the CL to use the newest version of the script. THe changes from 1.0.2 to 1.0.3 were all cosmetic - better word-wrapping of the commit message, etc.

However, your concern is very valid. Not all updates to the scripts are so innocuous. That's why one of the prerequisites in the README.md says:

"Please ensure that you have successfully synced your tree, and that your
toolchain is up-to-date" and then gives the `repo sync` command.

We got lucky this time.

Patch-set: 1