Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Raul, do you know what the EC_DIR variable is for (I seem to remember you playing around with it but I cant' remember for sure)
> 
> I'm guessing it's to catch the case where you call it from the sysroot instead of directly from the ec directory:
> 
> $ which flash_ec
> /usr/bin/flash_ec

Ah, that makes more sense. Raul, do happen to understand the intended behavior?

I'm guessing the idea is that if it is not sitting in the EC source directory, than the implicit default --image parameter will not work.
Maybe we can make this ec directory detection a bit more robust. Maybe check for "build" directory?

Patch-set: 1
Label: Commit-Queue=0