Update patch set 1
Patch Set 1:

(4 comments)

This patch appears to fix two more-or-less orthogonal bugs. 1) We don't actually send DP Exit Mode when we mean to, and 2) we somehow make it into dpm_attempt_mode_entry while discovery remains to be done. Can you split this up into separate CLs for each of those?

WRT bug 2) above, I don't think we should be able to make it into dpm_attempt_mode_entry when discovery remains to be done. If we are, that probably indicates a bug in pe_attempt_port_discovery or in how PE_FLAGS_DISCOVER_PORT_IDENTITY (misleadingly named) is handled. I haven't been able to reproduce this, and I don't immediately see such a bug. This approach seems like a good way to reduce pain for now.

Patch-set: 1