cr50/board: add board-local FIPS setting
Implement board-local configuraration of FIPS 140-2/3 policy as
complementary to FWMP policy. This is intended mostly for lab
testing and dogfooding, when FWMP policy is not feasible.

board_fips_enforced() returns status of FIPS from FWMP and NVRAM
and caches state to avoid expensive operations later.

BUG=b:138577491
TEST=manual, make buildall -j
Actual test command to be added in upcoming CLs

Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Change-Id: I8fa651e56e6e76a87bbc4dd911e7a8c0546e7e0f