Update patch set 1
Patch Set 1: -Code-Review

> Patch Set 1: Code-Review-2
> 
> Due to the CL chain [1] was merged, some configs / variables are moved from baseboard.h to each board.h. As a result, you would need to re-run the script for copying waddledee from newest version.
> 
> [1] https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2235238

new_variant.py doesn't have support to re-run just one of the 'create' or 'add' scripts. Even if it did, the scripts themselves don't support running again in-place to refresh everything.

The good news is that I can see how to recover. You can upload a newly-created version of the EC variant on top of this CL by following these steps:
1. `repo abandon` the create_drawcia_${DATE} branch on your local checkout
2. re-run create_initial_ec_image.sh
3. `git commit --amend` and replace the Change-Id with the value from this CL

If you have used new_variant.py to create the firmware for drawcia, you'll also need to edit ~/.new_variant.yaml and update the keys for 'ec_image' to use the branch name and CL number from the new CL. But you won't have to update the Change-Id since you copied it from this CL.

If it sounds complex and involved, that's because it is. When I was designing the program, I didn't foresee these kinds of cases occurring. If it happens again, I'm going to have to figure out a way to have the program handle it.

If you need more support, please contact me directly, and I can help you through the steps. It will be a learning experience for both of us.

Patch-set: 1
Label: Code-Review=0