usbc: Add config option to enable SS mux as UFP
This CL adds a new config option for superspeed mux (SS_MUX) which
will cause the USB mode of mux to be enabled by type-c state machine
only and not dependent on PD messaging. This is required for a type-c
dock which needs to expose USB3 hubs.

BUG=b:195042155
BRANCH=none
TEST=Verfied that when I connect USBC only source on gingerbread that
the TUSB1064 usb mux gets configured to enable USB3.1 mode.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I57de5dafd651a4c3d0ee23894195d9b44eeb0c3e