TCPMv2: Add Entry flow for Thunderbolt mode
Create a separate module for representing Thunderbolt mode which decides
wheather and how we can enter into Thunderbolt alternate mode.

Modify the Device Policy Manger (usb_pd_dpm.c) to first check if the
port partner and cable is a Thunderbolt Compatible. If not, continue the
rest of the PD flow undisturbed.

BUG=b:152235216
BRANCH=none
TEST=1. Attach Thubderbolt dock, DUT should enter Thunderbolt mode
     2. Attach Type-C dock, DUT should enter DP mode
     3. Attach DP dongle, DUT should enter DP mode

Change-Id: I369eb6337144676996c9d94a412eec060eacd273
Signed-off-by: Ayushee <ayushee.shah@intel.com>