tcpci: debug cleanup
DEBUG_GET_CC was already present to output CC
information on every get_cc call.  This CL
adds in spam reduction to not continue to
output the values if they are the same as
they were the last time they were output.

DEBUG_ROLE_CTRL_UPDATES was added to output
RoleControl register information every time
it is written in the TCPCI code.

BUG=none
BRANCH=none
TEST=make buildall

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Ic6eeadce8813908f1b89e4b756e90e95b1f9ca01