Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> I am a bit uncomfortable with this patch:
> 
> - it goes straight to NvDeleteEntity which is called by the rest of the TPM2 library after various verification steps.
> 
> - the TPM2_NV_UndefineSpace which also calls NvDeleteIntity even has a comment reading '// Indexes in the virtual range cannot be undefined.'
> 
> - we already have nvmem_erase_tpm_data, so with this patch there will be two different ways to delete objects - one operating on the flash storage, another one going through the TPM2 library shortcuts. I think it would be more consistent to modify nvmem_erase_tpm_data() to have a way to delete only a subset of objects if required.
> 
> Also, a general question: why would we want to disable FIPS mode, is there a problem with having it always on?

It does check that handle exist, and is supposed to be used only with 2 U2F objects, not a generic code. And it's not a virtual range, it's hidden.

I can't easily translate these TPM objects to objects used by new_nvmem, so just calling second part of nvmem_erase_tpm_data to compact and overwrite deleted objects.

And this code is called upon switching from non-FIPS to FIPS and back.  The first part is of interest.

Patch-set: 4
Label: Commit-Queue=0