Update patch set 1
Patch Set 1:

I thought it was said last time that if the EC reset, the Type-C circuitry would reset as well. Is this not the case? Well, actually it looks like I just said that with no confirmation. :p

> In my opinion, if the hardware is designed in such a way that an EC reset also resets the Type-C port, then during puff's board init, it should clear the BBRAM flag (checking the reset cause of course).

I agree with Peter that the reset cause needs to be checked.

Patch-set: 1