Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I thought it was said last time that if the EC reset, the Type-C circuitry would reset as well. Is this not the case? Well, actually it looks like I just said that with no confirmation. :p
> 
> > In my opinion, if the hardware is designed in such a way that an EC reset also resets the Type-C port, then during puff's board init, it should clear the BBRAM flag (checking the reset cause of course).
> 
> I agree with Peter that the reset cause needs to be checked.

So looking at the schematics, the TCPC and TCPPC are not connected to any of the system reset lines.
The TCPC has a reset line that the EC controls (and this is toggled as part of this function).
However, there doesn't appear to be any distinguishing between a power-on and the EC reset-pin causes.
the reset-pin cause is set whenever there is a power-on or an EC restart (like going into recovery mode).
BTW, this function is only called when the reset cause is reset-pin or power-on.
So what other reset cause should I be checking?

Further, if we clear this flag, then the sysjump_occurred flag is never set, so we end up with the same problem this is trying to fix - that the system thinks there is no existing connection, and gets hung up waiting for capabilities from the charger, causing a hard reset.

So resetting the TCPC doesn't seem to reset the connection, so sending a soft-reset seems
like the right thing to do here, so maybe I should rename sysjump_occurred to send_soft_reset
and set that from this function.
The interactions here between all these moving parts is doing my head in.

Patch-set: 1
Label: Commit-Queue=0