Update patch set 6
Patch Set 6:

Merged changes are up. There is still an issue with NCT3807 seeing the charger unless I change the pr_swap_complete for SRC->SNK to change the RoleControl register to only assert Rd on one line.  It can go back to both after this one time but for some reason they need only one to be asserted in order for ACOK to be driven by the TCPC

I am not marking this as verified but would appreciate comments so I can continue to update while I wait for Nuvoton to help me understand what is happening here.

Patch-set: 6