flash_ec: remove verify-whole option for npcx_(int_)spi
This patch removes the verfication of the whole firmware area for
npcx_spi and npcx_int_spi chip. The reason is flash_ec programs WP_RO
and EC_RW area only so it should verify partially on those area only,
not the whole area. The option "--fast-verify" already was used for that
purpose in flash_ec script.

BUG=b:158864362
BRANCH=none
TEST=ran flash_ec on nami.

Signed-off-by: Namyoon Woo <namyoon@google.com>
Change-Id: If08194697e1b15b13c592be94d5268359edd6b5a