Update patch set 1
Patch Set 1:

I like the path this is going much better than breaking the older hardware.

Can you break this into two CLs?  One that is really small that undefs CONFIG_POWER_SIGNAL_RUNTIME_CONFIG in config and includes power.h use of that. The second one will include you enabling it with all of the other changes to make that happen.  This will separate the introduction of the feature from you enabling it.

Patch-set: 1