TCPMv2: Exit Attached.SRC on open even during swap
If a port doesn't clear the PR swap in progress flag while switching
from a source to a sink, the port can end up permanently stuck in
Attached.SRC even when the port partner is unplugged.

This removes the check to the PR swap in progress flag when Open is
detected on the CC lines.  This follows the type-c spec's exit
conditions, and we would not expect to see CC open during a PR swap
since the source sets Rd before the new source will set Rp.

BRANCH=None
BUG=b:158613480
TEST=on mancomb, unplugging after failed PR swap with dock
allows port to go unattached again.  Normal PR swap process is able to
complete with dock plus servo_v4 and Apple 3-in-1

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Idc4c193597c6b4a791d18e38bc5111d0342e512c